Update README.md
This project is a frontend clone of the official PhonePe website's landing page, designed using HTML & CSS. The goal of this project was to replicate the sleek and modern UI of PhonePe while ensuring responsiveness across different devices.

With clean code structure, proper CSS styling, and a focus on mobile-friendly design, this clone serves as a great example of mastering frontend development skills. It showcases expertise in Flexbox, Grid, and responsive layouts.

Perfect for learning and demonstrating UI/UX implementation in real-world projects! 🚀